pyproject.toml
[tool.poetry]
name = "my-trading-bot"
version = "0.1.0"
description = ""
authors = ["leo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyupbit = "^0.2.34"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
